+++ presentation/src/commonMain/kotlin/com/moneyplusplus/presentation/feature/income/screens/addincome/AddIncomeContract.kt
@@ -1,9 +1,12 @@
 package com.moneyplusplus.presentation.feature.income.screens.addincome
 
+import com.moneyplusplus.presentation.base.UiEffect
+import com.moneyplusplus.presentation.base.UiIntent
 import com.moneyplusplus.domain.entity.Income
 import com.moneyplusplus.presentation.base.UiState
 import com.moneyplusplus.presentation.utils.currentDate
 import kotlinx.datetime.LocalDate
+
 
 data class AddIncomeState(
     val amount: Int? = null,
@@ -22,4 +25,22 @@ fun AddIncomeState.toIncome(): Income? {
         date = date,
         note = note
     )
+}
+
+sealed interface AddIncomeEffect: UiEffect {
+    data object ShowSuccess: AddIncomeEffect
+    data object ShowError: AddIncomeEffect
+    data object NavigateBack: AddIncomeEffect
+}
+
+sealed interface AddIncomeIntent: UiIntent {
+    data class SetAmount(val amount: Int?): AddIncomeIntent
+    data class SetNote(val note: String): AddIncomeIntent
+    data class SetDate(val date: LocalDate): AddIncomeIntent
+
+    data object OnDateClick: AddIncomeIntent
+    data object OnAddIncomeClick: AddIncomeIntent
+    data object OnBackClick: AddIncomeIntent
+
+    data object OnDatePickerDismiss: AddIncomeIntent
 }
