--- presentation/src/commonMain/kotlin/com/moneyplusplus/presentation/feature/income/screens/addincome/AddIncomeEffect.kt
@@ -1,9 +1,0 @@
-package com.moneyplusplus.presentation.feature.income.screens.addincome
-
-import com.moneyplusplus.presentation.base.UiEffect
-
-sealed interface AddIncomeEffect: UiEffect {
-    data object ShowSuccess: AddIncomeEffect
-    data object ShowError: AddIncomeEffect
-    data object NavigateBack: AddIncomeEffect
-}
